refactor: move controllers into their domain packages
Pure relocation, no behavior changes: the api/v1 flat controller list
dissolves into domain-cohesive packages (geo, logs, analytics, security,
auth, system, realtime), with controllers/ subpackages where a domain
has several. Route paths, names, and registration in server/routes.py
are unchanged apart from import paths; the /api/v1 URL prefix stays on
the controllers until the phase 4 router lands. The app's own log-file
controller lands in system/ (an admin feature over services/logfiles),
separate from the nginx access-log logs/ domain, and ingestion stats
join it for the same reason.
