+++ geometrikks/server/routes.py
@@ -9,19 +9,19 @@ from litestar.params import SkipValidation
 from litestar.response import File
 from litestar.types import ControllerRouterHandler
 
-from geometrikks.api.v1.geo_events_controller import GeoEventController
-from geometrikks.api.v1.geo_locations_controller import GeoLocationController
-from geometrikks.api.v1.access_log_controller import AccessLogController
-from geometrikks.api.v1.access_log_debug_controller import AccessLogDebugController
-from geometrikks.api.v1.analytics_controller import AnalyticsController
-from geometrikks.api.v1.auth_controller import AuthController
-from geometrikks.api.v1.crowdsec_controller import CrowdSecController
-from geometrikks.api.v1.logs_controller import LogsController
-from geometrikks.api.v1.system_controller import SystemController
-from geometrikks.api.v1.live_controller import crowdsec_feed, live_feed, logs_feed
-from geometrikks.api.v1.settings import read_settings
-from geometrikks.api.v1.stats import stats
-from geometrikks.api.health import health, health_ready
+from geometrikks.domain.geo.controllers.events import GeoEventController
+from geometrikks.domain.geo.controllers.locations import GeoLocationController
+from geometrikks.domain.logs.controllers.access_logs import AccessLogController
+from geometrikks.domain.logs.controllers.access_log_debug import AccessLogDebugController
+from geometrikks.domain.analytics.controllers import AnalyticsController
+from geometrikks.domain.auth.controllers import AuthController
+from geometrikks.domain.security.controllers import CrowdSecController
+from geometrikks.domain.system.controllers.logs import LogsController
+from geometrikks.domain.system.controllers.system import SystemController
+from geometrikks.domain.realtime.controllers import crowdsec_feed, live_feed, logs_feed
+from geometrikks.domain.system.controllers.settings import read_settings
+from geometrikks.domain.system.controllers.stats import stats
+from geometrikks.domain.system.controllers.health import health, health_ready
 from geometrikks.config.settings import Settings
